requirements.txt
Flask==2.3.2
pytesseract==0.3.10
opencv-python==4.8.0.76
Pillow==10.0.0
numpy==1.24.3